+++ scanpipe/pipes/d2d.py
@@ -313,13 +313,12 @@ def save_java_package_scan_results(codebase_resource, scan_results, scan_errors)
     Save the resource Java package scan results in the database as Resource.extra_data.
     Create project errors if any occurred during the scan.
     """
-    # note: we do not set a status on resources if we collected this correctly
+    # The status is only updated in case of errors.
     if scan_errors:
         codebase_resource.add_errors(scan_errors)
-        codebase_resource.status = flag.SCANNED_WITH_ERROR
+        codebase_resource.update(status=flag.SCANNED_WITH_ERROR)
     else:
-        codebase_resource.extra_data.update(scan_results)
-    codebase_resource.save()
+        codebase_resource.update_extra_data(scan_results)
 
 
 def _map_jar_to_source_resource(jar_resource, to_resources, from_resources):
@@ -412,8 +411,7 @@ def _map_path_resource(
     # Only create relations when the number of matches if inferior or equal to
     # the current number of path segment matched.
     if len(match.resource_ids) > match.matched_path_length:
-        to_resource.status = flag.TOO_MANY_MAPS
-        to_resource.save()
+        to_resource.update(status=flag.TOO_MANY_MAPS)
         return
 
     for resource_id in match.resource_ids: