+++ scanpipe/tests/pipes/test_scancode.py
@@ -411,8 +411,7 @@ class ScanPipeScancodePipesTest(TestCase):
         # Forcing package_uid for proper assertion with expected results
         package = project1.discoveredpackages.get()
         uuid = "ba110d49-b6f2-4c86-8d89-a6fd34838ca8"
-        package.package_uid = f"pkg:npm/is-npm@1.0.0?uuid={uuid}"
-        package.save()
+        package.update(package_uid=f"pkg:npm/is-npm@1.0.0?uuid={uuid}")
 
         scan_output_location = self.data_location / "is-npm-1.0.0_scan_package.json"
         summary = scancode.make_results_summary(project1, scan_output_location)