+++ CHANGELOG.rst
@@ -34,8 +34,14 @@ v33.0.0 (unreleased)
 - Improve the speed of Project ``reset`` and ``delete`` using the _raw_delete model API.
   https://github.com/nexB/scancode.io/issues/729
 
+- Specify ``update_fields`` during each ``save()`` related to Run tasks,
+  to force a SQL UPDATE in order to avoid any data loss when the model fields are
+  updated during the task execution.
+  https://github.com/nexB/scancode.io/issues/726
+
 - Add support for XLSX input in the ``load_inventory`` pipeline.
   https://github.com/nexB/scancode.io/issues/735
+
 
 v32.2.0 (2023-04-25)
 --------------------